Tutorial olayı eklendi

+++ core/src/main/java/com/muzo/ac2d/Trap.java
@@ -5,10 +5,9 @@ import com.badlogic.gdx.math.Vector2;
 public class Trap {
     public Vector2 position;
     public boolean isTriggered = false;
-    public float damage = 1f; // Düşmana verilen hasar
+    public int damage = 1;
 
     public Trap(float x, float y) {
-        // Tuzağın dünyadaki konumu
         this.position = new Vector2(x, y);
     }
 }
